requirements.txt
azure-common==1.1.25
azure-core==1.5.0
azure-cosmos==4.0.0
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.6
azure-functions==1.2.1
azure-nspkg==3.0.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
idna==2.9
pycparser==2.20
python-dateutil==2.8.1
requests==2.23.0
six==1.15.0
urllib3==1.25.9